+++ internal/cli/container.go
@@ -72,7 +72,7 @@ func NewContainer(version string) (*Container, error) {
 	}
 
 	// Set the default notebook if not already set
-	// might be overrided if --notebook-dir flag is present
+	// might be overridden if --notebook-dir flag is present
 	if osutil.GetOptEnv("ZK_NOTEBOOK_DIR").IsNull() && !config.Notebook.Dir.IsNull() {
 		notebookDir, err := paths.ExpandPath(config.Notebook.Dir.Unwrap())
 		if err != nil {